DevCard/requirements.txt
streamlit
qrcode
Pillow
numpy
cmake
pycairo